Refactor emitter classes more
