+++ Assets/Scripts/BulletManagement/PlayerEmitter.cs
@@ -3,12 +3,7 @@
 namespace BulletManagement
 {
     public class PlayerEmitter : EmitterBase
-    {
-        private void Start()
-        {
-            this.bulletPattern = this.GetComponent<BulletPattern>();
-        }
-        
+    { 
         // Update is called once per frame
         private void Update()
         {
@@ -20,18 +15,6 @@ namespace BulletManagement
                     EmitBullets();
                     timeCounter = 0f;
                 }
-                // EmitBullets();
-            }
-        }
-
-        protected override void EmitBullets()
-        {
-            for (int id = 0; id < this.emitBulletCount; ++id)
-            {
-                GameObject bulletGO = Instantiate(bulletPrefab, transform.position, Quaternion.identity);
-                Bullet bullet = bulletGO.GetComponent<Bullet>();
-                bullet.SetData(this, this.bulletPattern, id);
-                bullet.decayTime = bulletDecayTime;
             }
         }
     }
